refactored preemptive into a sepearte utils script so it can be used with other queues and can be invoke in a more appropriate place
