+++ src/sqs/sqs.ts
@@ -6,8 +6,9 @@ import { v4 as uuidv4 } from "uuid";
 import { statsd } from "config/statsd";
 import { Tags } from "hot-shots";
 import { sqsQueueMetrics } from "config/metric-names";
-import { stringFlag, StringFlags } from "config/feature-flags";
 import { ErrorHandler, ErrorHandlingResult, MessageHandler, QueueSettings, SQSContext, SQSMessageContext, SqsTimeoutError } from "~/src/sqs/sqs.types";
+import { stringFlag, StringFlags } from "config/feature-flags";
+import { preemptiveRateLimitCheck } from "utils/preemptive-rate-limit";
 
 //Maximum SQS Delay according to SQS docs https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-delay-queues.html
 const MAX_MESSAGE_DELAY_SEC: number = 15 * 60;
@@ -265,6 +266,10 @@ export class SqsQueue<MessagePayload> {
 		try {
 			const messageProcessingStartTime = Date.now();
 
+			if (await preemptiveRateLimitCheck(context, this.queueName, this.changeVisibilityTimeout)) {
+				//return;
+			}
+
 			// Change message visibility timeout to the max processing time
 			// plus EXTRA_VISIBILITY_TIMEOUT_DELAY to have some room for error handling in case of a timeout
 			await this.changeVisibilityTimeout(message, this.timeoutSec + EXTRA_VISIBILITY_TIMEOUT_DELAY, context.log);
@@ -327,7 +332,7 @@ export class SqsQueue<MessagePayload> {
 		return context.receiveCount >= this.maxAttempts;
 	}
 
-	public async changeVisibilityTimeout(message: Message, timeoutSec: number, logger: Logger): Promise<void> {
+	private async changeVisibilityTimeout(message: Message, timeoutSec: number, logger: Logger): Promise<void> {
 		if (!message.ReceiptHandle) {
 			logger.error(`No ReceiptHandle in message with ID = ${message.MessageId}`);
 			return;
@@ -349,6 +354,7 @@ export class SqsQueue<MessagePayload> {
 			VisibilityTimeout: Math.round(timeoutSec)
 		};
 		try {
+
 			await this.sqs.changeMessageVisibility(params).promise();
 		} catch (err) {
 			logger.error("Message visibility timeout change failed");
